Added controls for pinning and unpinning tabs, fixed pinned tabs in new windows
Expanded tab group views now show a pin icon by the favicon to indicate
pinned state.  This icon may be clicked to toggle the state, which will
be reflected when the tab is opened in either the active window, or in
a new window.

The pinned tab in new windows fix is a dirty kludge, but I don't know a
better way to do it with the callback being so unreliable.