Fixed opening a group in new window behavior
Prior to this update, tab groups opening in a new window had
inconsistent behavior in pinned tabs being pinned properly, or not all
tabs in the group opening.

Somehow, this was tied to the "focused" property on the
chrome.windows.create call being set to true.  Set to false, or with
the Chrome developer tools running, behavior is as expected.  For some
reason, the new window is focused anyway, possibly because of the
subsequent tab creation.